feat: replace old GRPC interface with Arrow Flight
