--- src/api/greptime/v1/common.proto
@@ -1,13 +1,0 @@
-syntax = "proto3";
-
-package greptime.v1;
-
-message RequestHeader {
-  string tenant = 1;
-}
-
-message ResultHeader {
-  uint32 version = 1;
-  uint32 code = 2;
-  string err_msg = 3;
-}
