--- src/api/greptime/v1/greptime.proto
@@ -1,19 +1,0 @@
-syntax = "proto3";
-
-package greptime.v1;
-
-import "greptime/v1/common.proto";
-import "greptime/v1/database.proto";
-
-service Greptime {
-  rpc Batch(BatchRequest) returns (BatchResponse) {}
-}
-
-message BatchRequest {
-  RequestHeader header = 1;
-  repeated DatabaseRequest databases = 2;
-}
-
-message BatchResponse {
-  repeated DatabaseResponse databases = 1;
-}
