+++ src/api/src/lib.rs
@@ -15,7 +15,6 @@
 pub mod error;
 pub mod helper;
 pub mod prometheus;
-pub mod result;
 pub mod serde;
 pub mod v1;
