+++ src/api/src/v1.rs
@@ -15,7 +15,5 @@
 #![allow(clippy::derive_partial_eq_without_eq)]
 tonic::include_proto!("greptime.v1");
 
-pub const GREPTIME_FD_SET: &[u8] = tonic::include_file_descriptor_set!("greptime_fd");
-
 mod column_def;
 pub mod meta;
