+++ src/catalog/Cargo.toml
@@ -21,7 +21,7 @@ common-time = { path = "../common/time" }
 datafusion.workspace = true
 datatypes = { path = "../datatypes" }
 futures = "0.3"
-futures-util = "0.3"
+futures-util.workspace = true
 lazy_static = "1.4"
 meta-client = { path = "../meta-client" }
 regex = "1.6"
