+++ src/common/error/src/lib.rs
@@ -24,6 +24,9 @@ pub mod prelude {
     pub use crate::ext::{BoxedError, ErrorExt};
     pub use crate::format::DebugFormat;
     pub use crate::status_code::StatusCode;
+
+    pub const INNER_ERROR_CODE: &str = "INNER_ERROR_CODE";
+    pub const INNER_ERROR_MSG: &str = "INNER_ERROR_MSG";
 }
 
 pub use snafu;
