+++ src/common/grpc/Cargo.toml
@@ -18,7 +18,7 @@ datafusion.workspace = true
 datatypes = { path = "../../datatypes" }
 flatbuffers = "22"
 futures = "0.3"
-prost = "0.11"
+prost.workspace = true
 snafu = { version = "0.7", features = ["backtraces"] }
 tokio.workspace = true
 tonic.workspace = true
