+++ src/datanode/src/error.rs
@@ -281,32 +281,8 @@ pub enum Error {
     #[snafu(display("Missing node id option in distributed mode"))]
     MissingMetasrvOpts { backtrace: Backtrace },
 
-    #[snafu(display("Invalid Flight ticket, source: {}", source))]
-    InvalidFlightTicket {
-        source: api::DecodeError,
-        backtrace: Backtrace,
-    },
-
     #[snafu(display("Missing required field: {}", name))]
     MissingRequiredField { name: String, backtrace: Backtrace },
-
-    #[snafu(display("Failed to poll recordbatch stream, source: {}", source))]
-    PollRecordbatchStream {
-        #[snafu(backtrace)]
-        source: common_recordbatch::error::Error,
-    },
-
-    #[snafu(display("Invalid FlightData, source: {}", source))]
-    InvalidFlightData {
-        #[snafu(backtrace)]
-        source: common_grpc::Error,
-    },
-
-    #[snafu(display("Failed to do Flight get, source: {}", source))]
-    FlightGet {
-        source: tonic::Status,
-        backtrace: Backtrace,
-    },
 }
 
 pub type Result<T> = std::result::Result<T, Error>;
@@ -336,8 +312,6 @@ impl ErrorExt for Error {
             | Error::CreateExprToRequest { source }
             | Error::InsertData { source } => source.status_code(),
 
-            Error::InvalidFlightData { source } => source.status_code(),
-
             Error::CreateSchema { source, .. }
             | Error::ConvertSchema { source, .. }
             | Error::VectorComputation { source } => source.status_code(),
@@ -362,8 +336,6 @@ impl ErrorExt for Error {
             | Error::RegisterSchema { .. }
             | Error::Catalog { .. }
             | Error::MissingRequiredField { .. }
-            | Error::FlightGet { .. }
-            | Error::InvalidFlightTicket { .. }
             | Error::IncorrectInternalState { .. } => StatusCode::Internal,
 
             Error::InitBackend { .. } => StatusCode::StorageUnavailable,
@@ -376,7 +348,6 @@ impl ErrorExt for Error {
             Error::BumpTableId { source, .. } => source.status_code(),
             Error::MissingNodeId { .. } => StatusCode::InvalidArguments,
             Error::MissingMetasrvOpts { .. } => StatusCode::InvalidArguments,
-            Error::PollRecordbatchStream { source } => source.status_code(),
         }
     }
