+++ src/frontend/Cargo.toml
@@ -6,7 +6,6 @@ license.workspace = true
 
 [dependencies]
 anymap = "1.0.0-beta.2"
-arrow-flight.workspace = true
 api = { path = "../api" }
 async-stream.workspace = true
 async-trait = "0.1"
@@ -28,12 +27,12 @@ datafusion-expr.workspace = true
 datanode = { path = "../datanode" }
 datatypes = { path = "../datatypes" }
 futures = "0.3"
-futures-util = "0.3"
+futures-util.workspace = true
 itertools = "0.10"
 meta-client = { path = "../meta-client" }
 moka = { version = "0.9", features = ["future"] }
 openmetrics-parser = "0.4"
-prost = "0.11"
+prost.workspace = true
 query = { path = "../query" }
 rustls = "0.20"
 serde = "1.0"
