+++ src/log-store/Cargo.toml
@@ -21,8 +21,8 @@ common-error = { path = "../common/error" }
 common-runtime = { path = "../common/runtime" }
 common-telemetry = { path = "../common/telemetry" }
 crc = "3.0"
-futures-util = "0.3"
 futures.workspace = true
+futures-util.workspace = true
 hex = "0.4"
 protobuf = { version = "2", features = ["bytes"] }
 raft-engine = "0.3"
