+++ src/meta-srv/Cargo.toml
@@ -25,7 +25,7 @@ h2 = "0.3"
 http-body = "0.4"
 lazy_static = "1.4"
 parking_lot = "0.12"
-prost = "0.11"
+prost.workspace = true
 regex = "1.6"
 serde = "1.0"
 serde_json = "1.0"
