+++ src/query/Cargo.toml
@@ -23,7 +23,7 @@ datafusion-physical-expr.workspace = true
 datafusion-sql.workspace = true
 datatypes = { path = "../datatypes" }
 futures = "0.3"
-futures-util = "0.3"
+futures-util.workspace = true
 metrics = "0.20"
 once_cell = "1.10"
 promql-parser = { git = "https://github.com/GreptimeTeam/promql-parser.git", rev = "d2f6ec4bbbae19b5156cfc977a6e7de9c6684651" }
