+++ src/servers/Cargo.toml
@@ -6,6 +6,7 @@ license.workspace = true
 
 [dependencies]
 aide = { version = "0.9", features = ["axum"] }
+arrow-flight.workspace = true
 api = { path = "../api" }
 async-trait = "0.1"
 axum = "0.6"
@@ -17,6 +18,7 @@ common-base = { path = "../common/base" }
 common-catalog = { path = "../common/catalog" }
 common-error = { path = "../common/error" }
 common-grpc = { path = "../common/grpc" }
+common-grpc-expr = { path = "../common/grpc-expr" }
 common-query = { path = "../common/query" }
 common-recordbatch = { path = "../common/recordbatch" }
 common-runtime = { path = "../common/runtime" }
@@ -36,7 +38,8 @@ once_cell = "1.16"
 openmetrics-parser = "0.4"
 opensrv-mysql = "0.3"
 pgwire = "0.6.3"
-prost = "0.11"
+pin-project = "1.0"
+prost.workspace = true
 query = { path = "../query" }
 rand = "0.8"
 regex = "1.6"
@@ -56,7 +59,6 @@ tokio.workspace = true
 tokio-rustls = "0.23"
 tokio-stream = { version = "0.1", features = ["net"] }
 tonic.workspace = true
-tonic-reflection = "0.5"
 tower = { version = "0.4", features = ["full"] }
 tower-http = { version = "0.3", features = ["full"] }
