+++ src/storage/Cargo.toml
@@ -19,13 +19,13 @@ common-telemetry = { path = "../common/telemetry" }
 common-time = { path = "../common/time" }
 datatypes = { path = "../datatypes" }
 futures.workspace = true
-futures-util = "0.3"
+futures-util.workspace = true
 lazy_static = "1.4"
 object-store = { path = "../object-store" }
 parquet = { workspace = true, features = ["async"] }
 paste.workspace = true
 planus = "0.2"
-prost = "0.11"
+prost.workspace = true
 regex = "1.5"
 serde.workspace = true
 serde_json = "1.0"
