+++ tests-integration/Cargo.toml
@@ -13,6 +13,7 @@ client = { path = "../src/client" }
 common-catalog = { path = "../src/common/catalog" }
 common-error = { path = "../src/common/error" }
 common-grpc = { path = "../src/common/grpc" }
+common-query = { path = "../src/common/query" }
 common-runtime = { path = "../src/common/runtime" }
 common-telemetry = { path = "../src/common/telemetry" }
 datanode = { path = "../src/datanode" }
