+++ tests/runner/Cargo.toml
@@ -9,5 +9,6 @@ async-trait = "0.1"
 client = { path = "../../src/client" }
 common-base = { path = "../../src/common/base" }
 common-grpc = { path = "../../src/common/grpc" }
+common-query = { path = "../../src/common/query" }
 sqlness = { git = "https://github.com/ceresdb/sqlness.git" }
 tokio.workspace = true
